+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Changelog
 
 ## [Unreleased]
+### Feature
+- `--quiet` flag
+- Source can now be file or directory
+- Accepts multiple sources
+- Improved logging configurability
+- Added support for csv comments
+
+### Breaking
+- Changed output dir to `--out`/`-o` option
+- Renamed `--pretty-print`/`-p` to `--minify`/`-m`
+- Renamed `--prefix-filename`/`-n` to `--prefix-identifiers`/`-p`
+- Prefix option now prefixes relative path to source
+- Missing translations are no longer skipped. Instead the default to ""
 
 ## [1.2.0] - 2021-03-17
 ### Added
@@ -17,7 +30,7 @@
 
 ## [1.0.1] - 2021-03-15
 ### Fixed
-- Tbump didn'T bump all version strings
+- Tbump didn't bump all version strings
 
 ## [1.0.0] - 2021-03-15
 ### Added